requirements.txt
# Core dependencies
langchain-openai==0.3.23
langchain>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.27.0
aiomysql>=0.2.0
aiosqlite>=0.19.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
flake8>=6.0.0  # Linting
mypy>=1.0.0   # Type checking 